refactor: move `RealDebridError` to `clients.errors`
