+++ cyberdrop_dl/managers/real_debrid/errors.py
@@ -1,15 +1,3 @@
-from __future__ import annotations
-
-from http import HTTPStatus
-from typing import TYPE_CHECKING
-
-from yarl import URL
-
-from cyberdrop_dl.clients.errors import CDLBaseError
-
-if TYPE_CHECKING:
-    from requests import Response
-
 ERROR_CODES = {
     -1: "Internal error",
     1: "Missing parameter",
@@ -49,27 +37,3 @@ ERROR_CODES = {
     35: "Infringing file",
     36: "Fair Usage Limit",
 }
-
-
-class RealDebridError(CDLBaseError):
-    """Base RealDebrid API error."""
-
-    def __init__(self, response: Response) -> None:
-        url = URL(response.url)
-        self.path = url.path
-        try:
-            JSONResp: dict = response.json()
-            code = JSONResp.get("error_code")
-            if code == 16:
-                code = 7
-            error = ERROR_CODES.get(code, "Unknown error")
-
-        except AttributeError:
-            code = response.status_code
-            error = f"{code} - {HTTPStatus(code).phrase}"
-
-        error = error.capitalize()
-
-        """This error will be thrown when a scrape fails."""
-        ui_message = f"{code} RealDebrid Error"
-        super().__init__(ui_message, message=error, status=code, origin=url)
